Bump numpy to 1.25.0

+++ skmultiflow/utils/validation.py
@@ -26,7 +26,9 @@ def check_random_state(seed):
         return np.random.RandomState(seed)
     if isinstance(seed, np.random.RandomState):
         return seed
-    raise ValueError('{} cannot be used to seed a numpy.random.RandomState instance'.format(seed))
+    raise ValueError(
+        "{} cannot be used to seed a numpy.random.RandomState instance".format(seed)
+    )
 
 
 def check_weights(weight, expand_length=1):
@@ -43,20 +45,24 @@ def check_weights(weight, expand_length=1):
         an np.array. If weight is not a single value, raises an error
 
     """
-    if isinstance(weight, (int, float, np.integer, np.float)):
+    if isinstance(weight, (int, float, np.integer, float)):
         if expand_length >= 1:
-            return np.array([weight] * expand_length, dtype=np.float)
+            return np.array([weight] * expand_length, dtype=float)
     elif isinstance(weight, list):
-        if all(isinstance(x, (int, float, np.integer, np.float)) for x in weight):
+        if all(isinstance(x, (int, float, np.integer, float)) for x in weight):
             if expand_length == 1:
                 return weight
     if isinstance(weight, np.ndarray):
-        if weight.size > 1 and all(isinstance(x, (int, float, np.integer, np.float)) for x in weight):
+        if weight.size > 1 and all(
+            isinstance(x, (int, float, np.integer, float)) for x in weight
+        ):
             if expand_length == 1:
                 return weight
-        elif weight.size == 1 and isinstance(weight[0], (int, float, np.integer, np.float)):
+        elif weight.size == 1 and isinstance(
+            weight[0], (int, float, np.integer, float)
+        ):
             if expand_length == 1:
                 return weight
             elif expand_length > 1:
-                return np.array([weight] * expand_length, dtype=np.float)
-    raise ValueError('Invalid weight(s): {}'.format(weight))
+                return np.array([weight] * expand_length, dtype=float)
+    raise ValueError("Invalid weight(s): {}".format(weight))
