+++ tests/utils/test_validation.py
@@ -28,15 +28,15 @@ def test_check_weights():
     weights = check_weights(float(test_value))
     assert weights == float(test_value)
 
-    weights = check_weights(np.float(test_value))
-    assert weights == np.float(test_value)
+    weights = check_weights(float(test_value))
+    assert weights == float(test_value)
 
-    weights = check_weights(np.int(test_value))
-    assert weights == np.int(test_value)
+    weights = check_weights(int(test_value))
+    assert weights == int(test_value)
 
-    weights = check_weights(np.array([test_value], np.float))
+    weights = check_weights(np.array([test_value], float))
     assert isinstance(weights, np.ndarray)
-    assert isinstance(weights[0], np.float)
+    assert isinstance(weights[0], float)
 
     weights = check_weights(np.array(np.arange(test_value)))
     assert isinstance(weights, np.ndarray)
@@ -54,10 +54,10 @@ def test_check_weights():
         assert w == int(test_value)
 
     with pytest.raises(ValueError):
-        check_weights([1.0, 2.0, 3.0, 4.0, 'invalid'])
+        check_weights([1.0, 2.0, 3.0, 4.0, "invalid"])
 
     with pytest.raises(ValueError):
-        check_weights('invalid')
+        check_weights("invalid")
 
     with pytest.raises(ValueError):
         check_weights([float(x) for x in range(test_value)], expand_length=10)
